Bug 1677426 - Part 3: Include JSFunction.h from Scope.h. r=nbp
Moved BindingKind and BindingLocation out of Scope.h,
added Scope.h -> JSFunction.h dependency,
and removed JSScript.h -> Scope.h dependency.

Differential Revision: https://phabricator.services.mozilla.com/D97662
