+++ js/src/vm/Interpreter.h
@@ -20,6 +20,7 @@
 
 namespace js {
 
+class WithScope;
 class EnvironmentIter;
 class PlainObject;
